Add patch to expose AU component type
We want to show "instrument" or "effect" in the build info
in the About window in VMPC2000XL. JUCE doesn't expose this
information for multi-component-type Audio Units, because
JUCE doesn't support multi-component-type Audio Units to
begin with.

To be sure, by multi-component-type we mean that
the Info.plist file contains not 1 AudioComponent entry, but
multiple. Apps like Koala, a sampler, expose themselves as
both an instrument (aumu) as well as effect (aumf), because
in some hosts there is no way to feed audio into an instrument.