Reinforce AUv3 component type exposure
We still use the same logic, except that it is applied on-demand,
rather than in the editor constructor. This means that the
AU component type will be derived when the user opens the About
window.

The reason for this change is that some hosts (Ableton Live) try
to spawn the AUv3's editor before spawning its processor.